Conflicts:
	rts/System/SpringApp.cpp

+++ rts/Game/LoadScreen.h
@@ -40,9 +40,9 @@ public:
 	void ResizeEvent();
 
 	/// Called when a key is released by the user
-	int KeyReleased(unsigned short k);
+	int KeyReleased(int k);
 	/// Called when the key is pressed by the user (can be called several times due to key repeat)
-	int KeyPressed(unsigned short k,bool isRepeat);
+	int KeyPressed(int k,bool isRepeat);
 
 
 private:
